only try to nuke the Pg database if it was created (like mysql)

+++ t/lib/Hailo/Test.pm
@@ -150,6 +150,7 @@ sub spawn_storage {
             if (system "createdb '$brainrs' >/dev/null 2>&1") {
                 $ok = 0;
             } else {
+                $self->{_created_pg} = 1;
                 # Kill Pg notices
                 $SIG{__WARN__} = sub { print STDERR @_ if $_[0] !~ m/NOTICE:\s*CREATE TABLE/; };
             }
@@ -178,8 +179,10 @@ sub unspawn_storage {
 
     given ($storage) {
         when (/Pg/) {
-            $nuke_db->();
-            system "dropdb '$brainrs'";
+            if ($self->{_created_pg}) {
+                $nuke_db->();
+                system "dropdb '$brainrs'";
+            }
         }
         when (/SQLite/) {
             $nuke_db->();
